requirements.txt
numpy==1.21.6
pandas==1.3.5
plotly==5.13.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
tenacity==8.1.0
nltk==3.8.1
wordcloud==1.8.2.2
matplotlib==3.1.3
FPDF==1.7.2
kaleido==0.2.1
Pillow==9.4.0